+++ python/ray/tune/stopper.py
@@ -1,7 +1,8 @@
 class Stopper:
-    """Abstract class for implementing a Tune experiment stopper.
+    """Base class for implementing a Tune experiment stopper.
 
-    Allows users to implement experiment-level stopping via ``stop_all``.
+    Allows users to implement experiment-level stopping via ``stop_all``. By
+    default, this class does not stop any trials.
 
     .. code-block:: python
 
@@ -10,17 +11,27 @@ class Stopper:
         from ray.tune import Stopper
 
         class TimeStopper(Stopper):
-                def __init__(self):
-                    self._start = time.time()
-                    self._deadline = 300
+            def __init__(self):
+                self._start = time.time()
+                self._deadline = 300
 
-                def stop_all(self):
-                    return time.time() - self._start > self.deadline
+            def stop_all(self):
+                return time.time() - self._start > self.deadline
 
         tune.run(Trainable, num_samples=200, stop=TimeStopper())
 
     """
 
+    def __call__(self, trial_id, result):
+        """Returns true if the trial should be terminated given the result."""
+        raise NotImplementedError
+
+    def stop_all(self):
+        """Returns true if the experiment should be terminated."""
+        raise NotImplementedError
+
+
+class NoopStopper(Stopper):
     def __call__(self, trial_id, result):
         return False
 
@@ -35,11 +46,14 @@ class FunctionStopper(Stopper):
     def __call__(self, trial_id, result):
         return self._fn(trial_id, result)
 
+    def stop_all(self):
+        return False
+
     @classmethod
     def is_valid_function(cls, fn):
-        is_valid = callable(fn) and not issubclass(type(fn), Stopper)
-        if hasattr(fn, "stop_all"):
+        is_function = callable(fn) and not issubclass(type(fn), Stopper)
+        if is_function and hasattr(fn, "stop_all"):
             raise ValueError(
                 "Stop object must be ray.tune.Stopper subclass to be detected "
                 "correctly.")
-        return is_valid
+        return is_function
