+++ python/ray/tune/tests/test_api.py
@@ -452,11 +452,11 @@ class TrainableFunctionApiTest(unittest.TestCase):
             for i in range(10):
                 reporter(test=i)
 
-        class Stopper:
+        class Stopclass:
             def stop(self, trial_id, result):
                 return result["test"] > 6
 
-        [trial] = tune.run(train, stop=Stopper().stop).trials
+        [trial] = tune.run(train, stop=Stopclass().stop).trials
         self.assertEqual(trial.last_result["training_iteration"], 8)
 
     def testStopper(self):
@@ -488,16 +488,16 @@ class TrainableFunctionApiTest(unittest.TestCase):
             for i in range(10):
                 reporter(test=i)
 
-        class Stopper:
+        class CustomStopper:
             def stop(self, result):
                 return result["test"] > 6
 
         def stop(result):
             return result["test"] > 6
 
-        with self.assertRaises(ValueError):
-            tune.run(train, stop=Stopper().stop)
-        with self.assertRaises(ValueError):
+        with self.assertRaises(TuneError):
+            tune.run(train, stop=CustomStopper().stop)
+        with self.assertRaises(TuneError):
             tune.run(train, stop=stop)
 
     def testEarlyReturn(self):
